- Lower memory footprint of permutation

+++ unittest/src/statistics/permutation_test.cpp
@@ -7,7 +7,7 @@
 
 namespace tfr {
 template <typename T>
-std::vector<uint64_t> get_histogram(const std::vector<T>& d) {
+std::vector<uint8_t> get_histogram(const std::vector<T>& d) {
 	auto s = create_stream_from_data("d", d, 0);
 	return get_permutation_histogram(ranged_stream(s, d.size()), 5);
 }
